Refactored Genetics class.
That thing was becoming a mess. Now it's split up:
- A class that implements a (pretty generic) GA.
- An evaluator that can also be used outside of the genetics.
- The Genetics class which only does UI work.

Subtle changes in behavior of the Genetics stuff may occur, and
there are plenty of things I noticed that could be improved.
For example, the "initial GA population" definitely behaves differently
now, and there's plenty of TODOs in the code.